deepen: inline enrichment-flags into summarize
`lib/ai/enrichment-flags.ts` (95 LOC, 3 exports) wrapped PostHog feature-flag
evaluation behind three thin functions — `isWhyItMattersEnabled`,
`isProviderEnabled`, `isEarningsMiniDeepDiveEnabled`. All three had exactly
one production caller: the enrichment gate in `summarizeFiling`. The
implementation was 50 LOC of "dynamic import posthog-node, call
getFeatureFlag, fail-safe to false" repeated three times with different
flag-key arguments. Tests asserted on private branches (env bypass,
PostHog error → false, unknown provider name → false) by mocking
posthog-server, then a separate `summarize-enrichment-gate.test.ts` mocked
the entire `enrichment-flags` module — a classic shallow seam that doubled
the surface tests had to cover.

Move all three functions into `summarize.ts` as private implementation
(`isWhyItMattersEnabled`, `isProviderEnabled`, `isEarningsMiniDeepDiveEnabled`)
plus a shared `evaluateEnrichmentFlag` helper. The dynamic posthog-server
import — required to keep posthog-node out of client bundles — moves with
them.

The deletion test passes: complexity concentrates beside the enrichment
gate where the flags are actually consulted. Bugs in flag-eval ordering,
fail-safe behaviour, or provider-name lookup now manifest right next to
the use site, not behind a module boundary.

Tests deleted: `__tests__/ai/enrichment-flags.test.ts` (10 direct tests
asserting on flag bypass / unknown-provider behaviour). Tests rewired:
`summarize-enrichment-gate.test.ts` now mocks `@/lib/analytics/posthog-server`
directly — the real external seam — and asserts on `getFeatureFlag` call
counts (with `expect.anything()` for the distinctId). The dead
`enrichment-flags` mock in `web-search-context.test.ts` is removed (the
module never imported from `enrichment-flags`).

Update the comment in `unified-prompts.ts` referencing the old module
path. The historical comment in `enrichment-spend.ts` ("Replaces the
process-local dailySpend accumulator in enrichment-flags.ts") stays — it
records project history, not a current dependency.

Dependency category: true external (PostHog). The deepened module takes
PostHog as an injected dependency via `getServerPostHog()`; tests provide
a mock adapter at that seam. The previous `enrichment-flags` interface
was a redundant indirection — one adapter (production) means a
hypothetical seam, not a real one.

Add `Enrichment Gate` term to CONTEXT.md.
